+++ src/main/resources/db/migration/V1__initial_version.sql
@@ -33,7 +33,7 @@ create table options (
     option_id          INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     unit_id            INT,
     name               TEXT NOT NULL,
-    points             INT,
+    points             DECIMAL ,
     pointsPerModel     BOOLEAN,
     parent_option      INT,
     constraint options_id_fk_1 foreign key (unit_id) references units (unit_id)
